# docs/swagger.yaml
basePath: /api
definitions:
...
  cardshttp.CardResponse:
    properties:
      description:
        type: string
      id:
        type: string
      is_archived:
        type: boolean
      list_id:
        type: string
      position:
        type: number
      priority:
        $ref: '#/definitions/entity.CardPriority'
      title:
        type: string
    type: object
  cardshttp.createCardRequest:
    properties:
      description:
        type: string
      list_id:
        type: string
      priority:
        $ref: '#/definitions/entity.CardPriority'
      title:
        type: string
    required:
    - list_id
    - title
    type: object
  cardshttp.moveCardRequest:
    properties:
      list_id:
        type: string
      position:
        type: number
    required:
    - position
    type: object
  cardshttp.updateCardRequest:
    properties:
      description:
        type: string
      is_archived:
        type: boolean
      position:
        type: number
      priority:
        $ref: '#/definitions/entity.CardPriority'
      title:
        type: string
    type: object
  entity.CardPriority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - PriorityLow
    - PriorityMedium
    - PriorityHigh
...
host: localhost:8003
info:
  contact: {}
  description: API for managing kanban boards
  title: Student Kanban API
  version: "1.0"
paths:
...
  /cards:
    get:
      description: Returns all cards belonging to a given list
      parameters:
      - description: List UUID
        in: query
        name: list_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cardshttp.CardResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all cards for a list
      tags:
      - cards
    post:
      consumes:
      - application/json
      description: Creates a new card inside a list
      parameters:
      - description: Card data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cardshttp.createCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cardshttp.CardResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new card
      tags:
      - cards
  /cards/{id}:
    delete:
      description: Deletes a card by its UUID
      parameters:
      - description: Card UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a card
      tags:
      - cards
    get:
      description: Returns a single card by its UUID
      parameters:
      - description: Card UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cardshttp.CardResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a card by ID
      tags:
      - cards
    put:
      consumes:
      - application/json
      description: Updates a card's fields. At least one field must be provided
      parameters:
      - description: Card UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cardshttp.updateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cardshttp.CardResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a card
      tags:
      - cards
  /cards/{id}/move:
    patch:
      consumes:
      - application/json
      description: Updates a card's position, optionally moving it to a different
        list
      parameters:
      - description: Card UUID
        in: path
        name: id
        required: true
        type: string
      - description: Move data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cardshttp.moveCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cardshttp.CardResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Move a card
      tags:
      - cards
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"